+++ cabal.project
@@ -9,6 +9,7 @@ packages:
   crucible-go/
   crucible-jvm/
   crucible-llvm/
+  crucible-mir/
   crucible-wasm/
   crucible-syntax/
   crucible-concurrency/
