+++ crucible-mir/README.md
@@ -1,0 +1,6 @@
+# crucible-mir
+
+This package implements a MIR frontend for Crucible. This exports all of the
+code that is useful across multiple projects that make use of MIR, such as
+[`crux-mir`](https://github.com/GaloisInc/crucible/blob/master/crux-mir) and
+[`crux-mir-comp`](https://github.com/GaloisInc/saw-script/blob/master/crux-mir-comp).
