+++ crucible-mir/crucible-mir.cabal
@@ -1,0 +1,56 @@
+name:                crucible-mir
+version:             0.1
+-- synopsis:
+-- description:
+homepage:            https://github.com/GaloisInc/crucible/blob/master/crucible-mir/README.md
+license:             BSD3
+license-file:        LICENSE
+author:              Joshua Gancher,
+                     Rob Dockins,
+                     Andrey Chudnov,
+                     Stephanie Weirich,
+                     Stuart Pernsteiner
+maintainer:          spernsteiner@galois.com
+copyright:           2017-2023 Galois, Inc.
+category:            Web
+build-type:          Simple
+cabal-version:       2.0
+extra-source-files:  README.md
+
+library
+  default-language: Haskell2010
+  build-depends: base >= 4.11 && < 5,
+                 aeson < 2.1,
+                 bv-sized,
+                 bytestring,
+                 prettyprinter >= 1.7.0,
+                 text,
+                 unordered-containers,
+                 crucible,
+                 parameterized-utils >= 1.0.8,
+                 containers,
+                 lens,
+                 vector,
+                 mtl,
+                 regex-compat,
+                 regex-base,
+                 transformers,
+                 what4,
+                 scientific       >= 0.3,
+                 template-haskell
+
+  hs-source-dirs: src
+  exposed-modules: Mir.JSON
+                   Mir.Generator
+                   Mir.Mir
+                   Mir.GenericOps
+                   Mir.ParseTranslate
+                   Mir.Pass
+                   Mir.Pass.AllocateEnum
+                   Mir.PP
+                   Mir.DefId
+                   Mir.FancyMuxTree
+                   Mir.Intrinsics
+                   Mir.TransTy
+                   Mir.Trans
+                   Mir.TransCustom
