+++ crux-mir/src/Mir/Generate.hs
@@ -10,25 +10,15 @@
 --
 -- This module sets up the process to compile the rust input file,
 -- extract the json representation, and parse as the MIR AST.
--- Also, entry point for translating MIR AST into crucible.
 -----------------------------------------------------------------------
 
 
-module Mir.Generate(generateMIR, translateMIR) where
+module Mir.Generate(generateMIR) where
 
-import Data.Foldable(fold)
-
-import Control.Lens hiding((<.>))
 import Control.Monad (when)
-import Control.Monad.ST
 
-import Control.Monad.IO.Class
-
-import qualified Data.Aeson as J
 import qualified Data.ByteString.Lazy as B
 import qualified Data.ByteString as BS
-import qualified Data.Map as M
-import qualified Data.Text as T
 
 import System.Environment
 import System.FilePath
@@ -40,23 +30,12 @@ import           Data.Time.Clock (UTCTime)
 
 import GHC.Stack
 
-import Prettyprinter (Pretty(..))
-
-import qualified Lang.Crucible.CFG.Core as C
-import qualified Lang.Crucible.FunctionHandle as C
-
 
 import Mir.Mir
-import Mir.JSON
-import Mir.Intrinsics(MIR)
+import Mir.ParseTranslate (parseMIR)
 import Mir.PP()
-import Mir.GenericOps (uninternTys)
-import Mir.Pass(rewriteCollection)
-import Mir.Generator(RustModule(..),CollectionState(..), rmCS, rmCFGs, collection)
-import Mir.Trans(transCollection, transStatics)
-import qualified Mir.TransCustom as Mir
 
-import Debug.Trace 
+import Debug.Trace
 
 
 getModificationTimeIfExists :: FilePath -> IO (Maybe UTCTime)
@@ -137,6 +116,7 @@ maybeLinkJson jsonFiles cacheFile = do
         B.readFile cacheFile
 
 
+libJsonFiles :: [FilePath]
 libJsonFiles =
     [ "libcore.mir"
     , "libcompiler_builtins.mir"
@@ -172,49 +152,11 @@ generateMIR inputFile keepRlib
     rlibsDir <- getRlibsDir
     let libJsonPaths = map (rlibsDir </>) libJsonFiles
     b <- maybeLinkJson (mirJsonOutFile rustFile : libJsonPaths) (linkOutFile rustFile)
-    parseMir (linkOutFile rustFile) b
+    parseMIR (linkOutFile rustFile) b
   | ext == ".json" = do
     b <- B.readFile inputFile
-    parseMir inputFile b
+    parseMIR inputFile b
   -- TODO: support directly reading .mir json+index format
   | otherwise = error $ show ext ++ " files are not supported"
   where
     (stem, ext) = splitExtension inputFile
-
-
-
-parseMir :: (HasCallStack, ?debug::Int) =>
-            FilePath
-         -> B.ByteString
-         -> IO Collection
-parseMir path f = do
-  let c = (J.eitherDecode f) :: Either String Collection
-  case c of
-      Left msg -> fail $ "JSON Decoding of " ++ path ++ " failed: " ++ msg
-      Right col -> do
-        when (?debug > 5) $ do
-          traceM "--------------------------------------------------------------"
-          traceM $ "Loaded module: " ++ path
-          traceM $ show (pretty col)
-          traceM "--------------------------------------------------------------"  
-        return $ uninternMir col
-
-uninternMir :: Collection -> Collection
-uninternMir col = uninternTys unintern (col { _namedTys = mempty })
-  where
-    -- NB: knot-tying is happening here.  Some values in `tyMap` depend on
-    -- other values.  This should be okay: the original `rustc::ty::Ty`s are
-    -- acyclic, so the entries in `tyMap` should be too.
-    tyMap = fmap (uninternTys unintern) (col^.namedTys)
-    unintern name = case M.lookup name tyMap of
-        Nothing -> error $ "missing " ++ show name ++ " in type map"
-        Just ty -> ty
-
-
--- | Translate a MIR collection to Crucible
-translateMIR :: (HasCallStack, ?debug::Int, ?assertFalseOnError::Bool, ?printCrucible::Bool) 
-   => CollectionState -> Collection -> C.HandleAllocator -> IO RustModule
-translateMIR lib col halloc =
-  let ?customOps = Mir.customOps in
-  let col0 = let ?mirLib  = lib^.collection in rewriteCollection col
-  in let ?libCS = lib in transCollection col0 halloc
